Wifi-HAL: WiFiLogger API’s check the logger feature capability
Check for the logger feature capabilities before processing the
logger API’s.Thus return WIFI_ERROR_NOT_SUPPORTED if the respective
feature capability is not set.
The current implementation only checks for this feature capability\
being 0, but could be extended in future, if the exact check of the
capability is required.

Change-Id: I4476f434c1cb26c67c06567ca154d5cca7ca338e
CRs-Fixed: 2150998